Remove ReportProvider type #5396
